requirements.txt
Flask==2.3.3
langchain==0.0.267
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.8.1
googletrans==4.0.0-rc1
matplotlib==3.7.2
python-dotenv==1.0.0
huggingface-hub==0.16.4
transformers==4.32.1
torch==2.2.0
numpy==1.24.3
pandas==2.0.3
pytest==7.4.0
gunicorn==21.2.0
boto3==1.28.36 